checkers/prototype/requirements.txt
gornilo
requests
urllib3
aiohttp>=3.9.0
Pillow>=10.0.0